+++ src/cli/test/ptbk.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env ts-node
 
 /**
- * Note: [🔺] Purpose of this file is to test the CLI in development environment
+ * Note: [🔺] Purpose of this file is to test and use the current CLI in development environment
  */
 
 import '../../_packages/cli.index.ts'; // <- Note: Register all the LLM providers, scrapers, etc. by importing this file