+++ src/pipeline/PipelineInterface/getPipelineInterface.ts
@@ -1,5 +1,5 @@
 import type { WritableDeep } from 'type-fest';
-import type { TODO_any } from '../../utils/organization/TODO_any';
+import { deepClone } from '../../utils/serialization/deepClone';
 import type { TODO_remove_as } from '../../utils/organization/TODO_remove_as';
 import { $deepFreeze } from '../../utils/serialization/$deepFreeze';
 import type { PipelineJson } from '../PipelineJson/PipelineJson';
@@ -13,9 +13,7 @@ import type { PipelineInterface } from './PipelineInterface';
  *
  * @public exported from `@promptbook/core`
  */
-export function getPipelineInterface(
-    pipeline: PipelineJson
-): PipelineInterface {
+export function getPipelineInterface(pipeline: PipelineJson): PipelineInterface {
     const pipelineInterface: WritableDeep<PipelineInterface> = {
         inputParameters: [],
         outputParameters: [],
@@ -25,11 +23,17 @@ export function getPipelineInterface(
         const { isInput, isOutput } = parameter;
 
         if (isInput) {
-            pipelineInterface.inputParameters.push(parameter as TODO_any);
+            pipelineInterface.inputParameters.push(
+                deepClone(parameter),
+                // <- Note: Clone to prevent mutation when `$deepFreeze` is called at the end
+            );
         }
 
         if (isOutput) {
-            pipelineInterface.outputParameters.push(parameter as TODO_any);
+            pipelineInterface.outputParameters.push(
+                deepClone(parameter),
+                // <- Note: Clone to prevent mutation when `$deepFreeze` is called at the end
+            );
         }
     }